complete exercises 4.18 to 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -1,5 +1,6 @@
 const blogRouter = require('express').Router()
 const logger = require('../utils/logger')
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
@@ -11,8 +12,10 @@ blogRouter.get('/', async (request, response) => {
 blogRouter.post('/', async (request, response, next) => {
     const body = request.body
 
-    const users = await User.find({})
-    const randomUser = users[Math.floor(Math.random() * users.length)]
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    const user = await User.findById(decodedToken.id)
 
     if (!body.likes) {
         body.likes = 0
@@ -23,15 +26,15 @@ blogRouter.post('/', async (request, response, next) => {
         author: body.author,
         url: body.url,
         likes: body.likes,
-        user: randomUser._id
+        user: user._id
     })
 
     try {
         const savedBlog = await blog.save()
         logger.info(`added ${blog.title} to the blog list`)
-        randomUser.blogs = randomUser.blogs.concat(savedBlog._id)
-        await randomUser.save()
-        logger.info(`blog linked to user ${randomUser.username}`)
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+        logger.info(`blog linked to user ${user.username}`)
         response.json(savedBlog.toJSON())
     } catch(exception) {
         next(exception)
@@ -39,11 +42,22 @@ blogRouter.post('/', async (request, response, next) => {
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
-    try {
-      await Blog.findByIdAndRemove(request.params.id)
-      response.status(204).end()
-    } catch (exception) {
-      next(exception)
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    const user = await User.findById(decodedToken.id)
+
+    const blogToDelete = await Blog.findById(request.params.id)
+
+    if ( blogToDelete.user._id.toString() === user._id.toString() ) {
+        try {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+          } catch (exception) {
+            next(exception)
+          }
+    } else {
+        return response.status(401).json({ error: `Unauthorized` })
     }
   })
 
@@ -54,19 +68,29 @@ blogRouter.put('/:id', async (request, response, next) => {
         body.likes = 0
     }
 
-    const blog = {
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes
-    }
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
 
-    try {
-        const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-        logger.info(`blog ${blog.title} successfully updated`)
-        response.json(updatedBlog.toJSON())
-    } catch (exception) {
-        next(exception)
+    const blogToUpdate = await Blog.findById(request.params.id)
+
+    if ( blogToUpdate.user._id.toString() === user._id.toString() ) {
+        const blog = {
+            title: body.title,
+            author: body.author,
+            url: body.url,
+            likes: body.likes
+        }
+
+        try {
+            const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+            logger.info(`blog ${blog.title} successfully updated`)
+            response.json(updatedBlog.toJSON())
+        } catch (exception) {
+            next(exception)
+        }
+    } else {
+        return response.status(401).json({ error: `Unauthorized` })
     }
 })
